Renderer refactor to make logic simpler
