+++ src/render/commonTypes.ts
@@ -1,0 +1,7 @@
+import * as BABYLON from 'babylonjs';
+
+export interface IEntityRenderData
+{
+    mesh: BABYLON.Mesh;
+    properties: Map<number, BABYLON.Mesh>;
+}
